+++ offer-manager-app/src/main/groovy/life/qbic/portal/offermanager/components/offer/create/CreateOfferViewModel.groovy
@@ -66,6 +66,8 @@ class CreateOfferViewModel {
     double overheads = 0
     @Bindable
     double totalPrice = 0
+    @Bindable
+    double totalDiscountAmount = 0
 
     @Bindable
     Boolean sequencingGridSelected
@@ -165,6 +167,7 @@ class CreateOfferViewModel {
         taxes = 0
         overheads = 0
         totalPrice = 0
+        totalDiscountAmount = 0
 
         projectTitleValid = false
         projectObjectiveValid = false